PYIC-7910 Move pages and contexts to config
